next/frontend: somehow get a translated frontend component to show up in next, but next translations are broken
