+++ src/packages/frontend/app/localize.tsx
@@ -7,13 +7,13 @@ import { ConfigProvider as AntdConfigProvider } from "antd";
 import type { Locale as AntdLocale } from "antd/lib/locale";
 import enUS from "antd/locale/en_US";
 import { isEmpty } from "lodash";
-import { createContext, useContext, useRef, useState } from "react";
+import { createContext, useContext, useState } from "react";
 import { IntlProvider } from "react-intl";
 import useAsyncEffect from "use-async-effect";
 
 type OnErrorFn = (typeof IntlProvider.defaultProps)["onError"];
 
-import { Loading, Paragraph, Text } from "@cocalc/frontend/components";
+import { Loading } from "@cocalc/frontend/components";
 import {
   DEFAULT_LOCALE,
   loadLocaleMessages,
@@ -24,6 +24,7 @@ import {
 } from "@cocalc/frontend/i18n";
 import { unreachable } from "@cocalc/util/misc";
 import { useAntdStyleProvider } from "./context";
+import { LOCALIZE_DEFAULT_ELEMENTS } from "./localize-default-elements";
 
 interface LanguageContextInterface {
   setLocale: (language: string) => void;
@@ -43,14 +44,6 @@ export function Localize({ children }: { children: React.ReactNode }) {
   const [antdLoc, setAntdLoc] = useState<AntdLocale | undefined>(undefined);
   const [messages, setMessages] = useState<Messages | undefined>(undefined);
   const { antdTheme } = useAntdStyleProvider();
-  const uniqueKey = useRef<{ [tag: string]: number }>({});
-
-  // Note: this is e.g. necessary to render text in a modal, where some caching happens, apparently
-  function getKey(tag: string): number {
-    const n = (uniqueKey.current[tag] ?? 0) + 1;
-    uniqueKey.current[tag] = n;
-    return n;
-  }
 
   useAsyncEffect(async () => {
     setMessages(await loadLocaleMessages(locale));
@@ -99,32 +92,7 @@ export function Localize({ children }: { children: React.ReactNode }) {
           messages={messages}
           defaultLocale={DEFAULT_LOCALE}
           onError={onError}
-          defaultRichTextElements={{
-            strong: (ch) => (
-              <Text strong key={getKey("strong")}>
-                {ch}
-              </Text>
-            ),
-            b: (ch) => (
-              <Text strong key={getKey("b")}>
-                {ch}
-              </Text>
-            ),
-            i: (ch) => (
-              <Text italic key={getKey("i")}>
-                {ch}
-              </Text>
-            ),
-            p: (ch) => <Paragraph key={getKey("p")}>{ch}</Paragraph>,
-            code: (ch) => (
-              <Text code key={getKey("code")}>
-                {ch}
-              </Text>
-            ),
-            ul: (e) => <ul key={getKey("ul")}>{e}</ul>,
-            ol: (e) => <ol key={getKey("ol")}>{e}</ol>,
-            li: (e) => <li key={getKey("li")}>{e}</li>,
-          }}
+          defaultRichTextElements={LOCALIZE_DEFAULT_ELEMENTS}
         >
           {renderApp()}
         </IntlProvider>
