+++ src/packages/frontend/client/project.ts
@@ -12,7 +12,8 @@ import { join } from "path";
 import { redux } from "@cocalc/frontend/app-framework";
 import computeServers from "@cocalc/frontend/compute/manager";
 import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
-import { dialogs, getIntl } from "@cocalc/frontend/i18n";
+import { dialogs } from "@cocalc/frontend/i18n";
+import { getIntl } from "@cocalc/frontend/i18n/get-intl";
 import { ipywidgetsGetBufferUrl } from "@cocalc/frontend/jupyter/server-urls";
 import { allow_project_to_run } from "@cocalc/frontend/project/client-side-throttle";
 import { ensure_project_running } from "@cocalc/frontend/project/project-start-warning";
