+++ src/packages/frontend/project_actions.ts
@@ -11,24 +11,7 @@ import { callback } from "awaiting";
 import { List, Map, Set, fromJS } from "immutable";
 import { isEqual } from "lodash";
 import { join } from "path";
-import type { ChatState } from "@cocalc/frontend/chat/chat-indicator";
-import { initChat } from "@cocalc/frontend/chat/register";
-import * as computeServers from "@cocalc/frontend/compute/compute-servers-table";
-import { modalParams } from "@cocalc/frontend/compute/select-server-for-file";
-import { TabName, setServerTab } from "@cocalc/frontend/compute/tab";
-import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
-import { chatFile } from "@cocalc/frontend/frame-editors/generic/chat";
-import { dialogs, getIntl } from "@cocalc/frontend/i18n";
-import { set_local_storage } from "@cocalc/frontend/misc";
-import Fragment, { FragmentId } from "@cocalc/frontend/misc/fragment-id";
-import fetchDirectoryListing from "@cocalc/frontend/project/fetch-directory-listing";
-import track from "@cocalc/frontend/user-tracking";
-import { retry_until_success } from "@cocalc/util/async-utils";
-import { DEFAULT_NEW_FILENAMES, NEW_FILENAMES } from "@cocalc/util/db-schema";
-import * as misc from "@cocalc/util/misc";
-import { reduxNameToProjectId } from "@cocalc/util/redux/name";
-import { MARKERS } from "@cocalc/util/sagews";
-import { client_db } from "@cocalc/util/schema";
+
 import { default_filename } from "@cocalc/frontend/account";
 import { alert_message } from "@cocalc/frontend/alerts";
 import {
@@ -36,21 +19,32 @@ import {
   project_redux_name,
   redux,
 } from "@cocalc/frontend/app-framework";
+import type { ChatState } from "@cocalc/frontend/chat/chat-indicator";
+import { initChat } from "@cocalc/frontend/chat/register";
 import { IconName } from "@cocalc/frontend/components";
+import * as computeServers from "@cocalc/frontend/compute/compute-servers-table";
+import { modalParams } from "@cocalc/frontend/compute/select-server-for-file";
+import { TabName, setServerTab } from "@cocalc/frontend/compute/tab";
+import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
 import { local_storage } from "@cocalc/frontend/editor-local-storage";
-import { get_editor } from "./editors/react-wrapper";
+import { chatFile } from "@cocalc/frontend/frame-editors/generic/chat";
 import {
   query as client_query,
   exec,
 } from "@cocalc/frontend/frame-editors/generic/client";
 import { set_url } from "@cocalc/frontend/history";
+import { dialogs } from "@cocalc/frontend/i18n";
+import { getIntl } from "@cocalc/frontend/i18n/get-intl";
 import {
   download_file,
   open_new_tab,
   open_popup_window,
+  set_local_storage,
 } from "@cocalc/frontend/misc";
+import Fragment, { FragmentId } from "@cocalc/frontend/misc/fragment-id";
 import * as project_file from "@cocalc/frontend/project-file";
 import { delete_files } from "@cocalc/frontend/project/delete-files";
+import fetchDirectoryListing from "@cocalc/frontend/project/fetch-directory-listing";
 import {
   ProjectEvent,
   SoftwareEnvironmentEvent,
@@ -100,7 +94,15 @@ import {
   ProjectStore,
   ProjectStoreState,
 } from "@cocalc/frontend/project_store";
+import track from "@cocalc/frontend/user-tracking";
 import { webapp_client } from "@cocalc/frontend/webapp-client";
+import { retry_until_success } from "@cocalc/util/async-utils";
+import { DEFAULT_NEW_FILENAMES, NEW_FILENAMES } from "@cocalc/util/db-schema";
+import * as misc from "@cocalc/util/misc";
+import { reduxNameToProjectId } from "@cocalc/util/redux/name";
+import { MARKERS } from "@cocalc/util/sagews";
+import { client_db } from "@cocalc/util/schema";
+import { get_editor } from "./editors/react-wrapper";
 
 const { defaults, required } = misc;
 
@@ -1375,8 +1377,9 @@ export class ProjectActions extends Actions<ProjectStoreState> {
     const computeServerAssociations =
       webapp_client.project_client.computeServers(this.project_id);
     const sidePath = chatFile(path);
-    const currentId =
-      await computeServerAssociations.getServerIdForPath(sidePath);
+    const currentId = await computeServerAssociations.getServerIdForPath(
+      sidePath,
+    );
     if (currentId != null) {
       // already set
       return;
@@ -2217,8 +2220,8 @@ export class ProjectActions extends Actions<ProjectStoreState> {
             dest_compute_server_id: opts.dest_compute_server_id,
           }
         : opts.src_compute_server_id
-          ? { compute_server_id: opts.src_compute_server_id }
-          : undefined),
+        ? { compute_server_id: opts.src_compute_server_id }
+        : undefined),
     });
 
     if (opts.only_contents) {
