+++ src/packages/next/package.json
@@ -90,6 +90,7 @@
     "react-dom": "^18.3.1",
     "react-google-recaptcha": "^2.1.0",
     "react-google-recaptcha-v3": "^1.9.7",
+    "react-intl": "^6.7.0",
     "serve-index": "^1.9.1",
     "sha1": "^1.1.1",
     "sharp": "^0.32.6",
